+++ C#/TollFeeCalculator.Sweden/Taxation.cs
@@ -28,14 +28,14 @@ namespace TollFeeCalculator.Sweden
 		};
 
 		public IEnumerable<TaxationTimespan> TaxationTimespans { get; } = null;
-		public float TimeSpanFee(DateTime dateTime)
+		public int TimeSpanFee(DateTime dateTime)
 		{
 			throw new NotImplementedException("Implemented in city specific class.");
 		}
 
-		public float MaxDailyFee { get; } = 0.0f;
+		public int MaxDailyFee { get; } = 0;
 		public int SingleChargeRuleMinutes { get; } = -1;
-		public float FeeForPassages(IVehicle vehicle, IEnumerable<DateTime> passages)
+		public int FeeForPassages(IVehicle vehicle, IEnumerable<DateTime> passages)
 		{
 			throw new NotImplementedException("Implemented in city specific class.");
 		}
